bootparam: refactor boot param with reference to kernel source
Lack of e820 entry setup was causing memory shortages in the fedora 31 kernel
and initrd loading. The offset in e820_map[E820MAX] was in an unintended
position, so I adjusted the padding and added unit tests. Also, when booting
from protected mode, all parameters except SetupHeader (offset 0x1f1) had to
be rebuilt from zero filling. With this fix, it is now possible to launch the
init process using the fedora 31 kernel and initrd.

Signed-off-by: Nobuhiro MIKI <nob@bobuhiro11.net>