Add amd64 bit to machine
This change allows callers of setup to specify
amd64.

This includes initial page tables and long
mode setup so we can start kernel at the
64-bit entry point.

It also includes support for KVM_TRANSLATE, which
is incredibly useful for debug.

golangci-lint is making a lot of bogus complaints.

Add a memsize flag, -m

usage
-m number -- default units are gigabytes

-m number[gGmMkK]

units of GiB, MiB, or KiB

tested by:

./gokvm -k vmlinux-5.10 -i /dev/null -t ""
./gokvm -k bzImage-5.10 -i /dev/null -t ""

both work.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>