kvm/: various fixes, preparing for ELF and 64-bit support
This is the first step in adding ELF file and 64-bit support.

Add document comments to kvm.go as suggested by several linters

add ioctl support helpers to generate ioctl commands.
This reduces the number of magic constants

Show code for KVM_TRANSLATE. This is incredibly helpful
when debugging VM startup, to find mistakes in page tables.
It is not ready to use until we bring in 64-bit support,
and leaving it in hurts the coverall test, so it is commented
out.

Add an ExitType and a stringer for it.

The stringer was generated by the Go stringer, which still has
lots of bugs around use of the errors package, necessitating a
workaround in how it was generated. Hence the comment in kvm.go
about it.

Use the stringer in machine/machine.go so users can see an
error string, not a magic number.

btw, some of the changes golangci-lint are either
- considered to be coding anti-patterns in Google (requiring line breaks)
- wrong, as in thinking all tests should be parallel

TestAddNum is now run in single step mode, to verify that SingleStep works.
If the test runs, but there are no recorded ExitDebug exits, that is
considered an error.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>