Add infrastructure for tracing and printing assembly
Now, if gokvm is invoked with -T, you will see traces
like this:

0x29f3edf:lea 0x1(%rcx),%rsi
0x29f3ee3:mov %r9,0x10(%rax)
0x29f3ee7:movzbl (%r8,%rcx,1),%ecx
0x29f3eec:mov %cl,(%r8,%rdx,1)
0x29f3ef0:cmp %rsi,0x28(%rax)
0x29f3ef4:cmove %r10,%rsi
0x29f3ef8:mov 0x10(%rax),%rdx
0x29f3efc:mov %rsi,%rcx
0x29f3eff:sub $0x1,%edi

Much more improvement is possible, but this is a useful first step.

What would be nice: if we had a simple signal handler in gokvm,
we could use it to turn tracing off and on.

E.g., we could start gokvm, and from another window,
killall -HUP gokvm

And it would flip the tracing flag.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>