do not start goroutines in packages; use errgroup
Starting goroutines in packages is now frowned upon.
Also, we can use errgroup to simplify the code.

This is not quite bringing them up to the top;
but they are in vmm, and they do use errgroup.

Now, when you do
^Ax
it does not call os.Exit(0), which is better behavior.

But we need to add context support, so we can cancel everything
when either a processor or the console exits.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>